Migrate game engine module to TypeScript
